Update: added database
Implemented sqlite database. It is called on for the internship, events, and research viewing button clicks. May implement a jobs one for it in the future. Currently displays information in a listview, although this could optimized with a recylerview and an expandable view list to allow users to view descriptions. However, this is a final project, and I don't know if it will ever see the light of day beyond this class. Our project is already beyond the minimum requirements so I'm not sweating it. Also fixed a broken link to the acm website.